requirements.txt
torch>=2.1.0
torchvision>=0.16.2
torchaudio
isort
black
flake8
einops>=0.7.0
gradio
tokenizers
wandb
deepspeed
peft
bitsandbytes
scikit-learn
requests
transformers==4.39.3
pillow>=10.0.1
seaborn>=0.13.0
timm>=0.9.12
accelerate>=0.25.0
trl
omegaconf
datasets
pre-commit
opencv-python
chardet
torchmetrics
statsmodels
nltk
ninja
natsort
pydicom
scikit-image
openai
open_clip_torch
fast-bleu
bert-score
python-dotenv
f1chexbert==0.0.2
pycocoevalcap
# # faiss-gpu depending on CUDA version
# faiss-gpu-cu11
faiss-gpu-cu12
albumentations
pytorch-lightning
ipdb
segmentation-models-pytorch